ssg-md tests: partial adaptation — 151/178 test files compile clean
MutableDataSet.set extension for JavaCollection/java.util.Collection
bridging. Method parens, getter renames, Nullable removal, and type
fixes across 129 test files. 27 test files still have compile errors
(text visibility, NodeVisitor constructor, remaining type mismatches).

+++ ssg-md/src/test/scala/ssg/md/core/test/ComboKramdownCompatibilitySpecTest.scala
@@ -41,7 +41,7 @@ object ComboKramdownCompatibilitySpecTest {
   val SPEC_RESOURCE:     String           = "/ssg/md/core/test/core_kramdown_compatibility_spec.md"
   val RESOURCE_LOCATION: ResourceLocation = ResourceLocation.of(classOf[ComboKramdownCompatibilitySpecTest], SPEC_RESOURCE)
 
-  val OPTIONS: DataHolder = new MutableDataSet().setFrom(ParserEmulationProfile.KRAMDOWN).set(HtmlRenderer.INDENT_SIZE, 4).toImmutable()
+  val OPTIONS: DataHolder = new MutableDataSet().setFrom(ParserEmulationProfile.KRAMDOWN).set(HtmlRenderer.INDENT_SIZE, java.lang.Integer.valueOf(4)).toImmutable()
 
   /** Merged: CoreRendererOptions.BASE_OPTIONS + this test's OPTIONS */
   val MERGED_OPTIONS: DataHolder =
